Corrigé l'erreur "int cannot be dereferenced" dans MetricsTracker.java

+++ src/main/java/com/project/common/utils/MetricsTracker.java
@@ -273,7 +273,7 @@ public class MetricsTracker extends BaseTrainingListener {
                 for (Integer classIdx : lastMetrics.getPerClassMetrics().keySet()) {
                     ClassMetrics classMetrics = lastMetrics.getClassMetrics(classIdx);
                     writer.write(String.format("%d,%.6f,%.6f,%.6f\n",
-                            classIdx, classMetrics.getPrecision(),
+                            classIdx.intValue(), classMetrics.getPrecision(),
                             classMetrics.getRecall(), classMetrics.getF1Score()));
                 }
             }
